chore: Updated to go 1.23
This update moves us to go 1.23. Sadly the astar library that we used introduced a breaking change, so we needed to convert the neighbors function to use an iter.Seq instead of returning an array.

This is a nicer way to implement the code through.

Also introducting a .gitignore which apparently I hadn't put in before this.